Extract Batch 5 query/mutation tool handlers into src/tools/
Moves recent_memories, memory_graph, project_memory_summary, sync,
move_memory, relate, unrelate, and consolidate handlers from
src/index.ts into dedicated modules under src/tools/.

Each module exports registerXxxTool(server, ctx) following the
established ctx-based dependency injection pattern.

Also extracts warnAboutPendingMigrationsOnStartup into
src/startup.ts so it remains callable from index.ts.
