+++ src/index.ts
@@ -53,6 +53,14 @@ import { registerGetTool } from "./tools/get.js";
 import { registerWhereIsMemoryTool } from "./tools/where-is-memory.js";
 import { registerListTool } from "./tools/list.js";
 import { registerDiscoverTagsTool } from "./tools/discover-tags.js";
+import { registerRecentMemoriesTool } from "./tools/recent-memories.js";
+import { registerMemoryGraphTool } from "./tools/memory-graph.js";
+import { registerProjectMemorySummaryTool } from "./tools/project-memory-summary.js";
+import { registerSyncTool } from "./tools/sync.js";
+import { registerMoveMemoryTool } from "./tools/move-memory.js";
+import { registerRelateTool } from "./tools/relate.js";
+import { registerUnrelateTool } from "./tools/unrelate.js";
+import { registerConsolidateTool } from "./tools/consolidate.js";
 import { suggestAutoRelationships } from "./auto-relate.js";
 import {
   computeHybridScore,
@@ -632,1538 +640,21 @@ registerListTool(server, ctx);
 
 registerDiscoverTagsTool(server, ctx);
 
-// ── recent_memories ───────────────────────────────────────────────────────────
-server.registerTool(
-  "recent_memories",
-  {
-    title: "Recent Memories",
-    description:
-      "Show the most recently updated memories.\n\n" +
-      "Use this when:\n" +
-      "- You want to see what changed most recently\n" +
-      "- You are resuming work and want a quick chronological view\n\n" +
-      "Do not use this when:\n" +
-      "- You need topic-based search; use `recall`\n" +
-      "- You need a tag/scope inventory; use `list`\n\n" +
-      "Returns:\n" +
-      "- Recently updated memories with ids, timestamps, storage labels, and basic metadata\n\n" +
-      "Read-only.\n\n" +
-      "Typical next step:\n" +
-      "- Use `get` for exact inspection or `update` to continue refining a recent note.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      cwd: projectParam,
-      scope: z.enum(["project", "global", "all"]).optional().default("all"),
-      storedIn: z.enum(["project-vault", "main-vault", "any"]).optional().default("any"),
-      limit: z.number().int().min(1).max(20).optional().default(5),
-      includePreview: z.boolean().optional().default(true),
-      includeStorage: z.boolean().optional().default(true),
-      lifecycle: z.enum(["temporary", "permanent"]).optional().describe("Filter results by lifecycle. Useful for recovering working-state with `lifecycle: temporary` after `project_memory_summary` orientation."),
-    }),
-    outputSchema: RecentResultSchema,
-  },
-  async ({ cwd, scope, storedIn, limit, includePreview, includeStorage, lifecycle }) => {
-    await ensureBranchSynced(cwd);
+registerRecentMemoriesTool(server, ctx);
 
-    const { project, entries } = await collectVisibleNotes(cwd, scope, undefined, storedIn);
-    let filteredEntries = entries;
-    if (lifecycle) {
-      filteredEntries = entries.filter(({ note }) => note.lifecycle === lifecycle);
-    }
-    const recent = [...filteredEntries]
-      .sort((a, b) => b.note.updatedAt.localeCompare(a.note.updatedAt))
-      .slice(0, limit);
+registerMemoryGraphTool(server, ctx);
 
-    if (recent.length === 0) {
-      const structuredContent: RecentResult = { action: "recent_shown", project: toProjectRef(project), count: 0, limit: limit || 5, notes: [] };
-      return { content: [{ type: "text", text: "No memories found." }], structuredContent };
-    }
+registerProjectMemorySummaryTool(server, ctx);
 
-    const header = project && scope !== "global"
-      ? `Recent memories for ${project.name}:`
-      : "Recent memories:";
-    const lines = recent.map((entry) => formatListEntry(entry, {
-      includePreview,
-      includeStorage,
-      includeUpdated: true,
-    }));
-    
-    const textContent = `${header}\n\n${lines.join("\n")}`;
-    
-      const structuredNotes = recent.map(({ note, vault }) => ({
-        id: note.id,
-        title: note.title,
-        project: noteProjectRef(note),
-        tags: note.tags,
-        lifecycle: note.lifecycle,
-        vault: storageLabel(vault),
-        updatedAt: note.updatedAt,
-        preview: includePreview && note.content ? note.content.substring(0, 100) + (note.content.length > 100 ? "..." : "") : undefined,
-    }));
-    
-    const structuredContent: RecentResult = {
-      action: "recent_shown",
-      project: toProjectRef(project),
-      count: recent.length,
-      limit: limit || 5,
-      notes: structuredNotes,
-    };
-    
-    return { content: [{ type: "text", text: textContent }], structuredContent };
-  }
-);
+registerSyncTool(server, ctx);
 
-// ── memory_graph ──────────────────────────────────────────────────────────────
-server.registerTool(
-  "memory_graph",
-  {
-    title: "Memory Graph",
-    description:
-      "Show memory nodes and their relationships as a graph-oriented view.\n\n" +
-      "Use this when:\n" +
-      "- You want to inspect how notes connect across a topic or project\n" +
-      "- You are evaluating whether relationships are too sparse or too dense\n\n" +
-      "Do not use this when:\n" +
-      "- You only need one note; use `get`\n" +
-      "- You only need ranked topic matches; use `recall`\n\n" +
-      "Returns:\n" +
-      "- Memory nodes and relationship edges for the requested slice\n\n" +
-      "Read-only.\n\n" +
-      "Typical next step:\n" +
-      "- Use `get`, `relate`, `unrelate`, or `consolidate` based on what the graph reveals.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      cwd: projectParam,
-      scope: z.enum(["project", "global", "all"]).optional().default("all"),
-      storedIn: z.enum(["project-vault", "main-vault", "any"]).optional().default("any"),
-      limit: z.number().int().min(1).max(50).optional().default(25),
-    }),
-    outputSchema: MemoryGraphResultSchema,
-  },
-  async ({ cwd, scope, storedIn, limit }) => {
-    await ensureBranchSynced(cwd);
+registerMoveMemoryTool(server, ctx);
 
-    const { project, entries } = await collectVisibleNotes(cwd, scope, undefined, storedIn);
-    if (entries.length === 0) {
-      const structuredContent: MemoryGraphResult = { action: "graph_shown", project: toProjectRef(project), nodes: [], limit, truncated: false };
-      return { content: [{ type: "text", text: "No memories found." }], structuredContent };
-    }
+registerRelateTool(server, ctx);
 
-    const visibleIds = new Set(entries.map((entry) => entry.note.id));
-    const lines = entries
-      .filter((entry) => (entry.note.relatedTo?.length ?? 0) > 0)
-      .slice(0, limit)
-      .map((entry) => {
-        const edges = (entry.note.relatedTo ?? [])
-          .filter((rel) => visibleIds.has(rel.id))
-          .map((rel) => `${rel.id} (${rel.type})`);
-        return edges.length > 0 ? `- ${entry.note.id} -> ${edges.join(", ")}` : null;
-      })
-      .filter(Boolean);
+registerUnrelateTool(server, ctx);
 
-    if (lines.length === 0) {
-      const structuredContent: MemoryGraphResult = { action: "graph_shown", project: toProjectRef(project), nodes: [], limit, truncated: false };
-      return { content: [{ type: "text", text: "No relationships found for that scope." }], structuredContent };
-    }
-
-    const header = project && scope !== "global"
-      ? `Memory graph for ${project.name}:`
-      : "Memory graph:";
-    
-    const textContent = `${header}\n\n${lines.join("\n")}`;
-    
-    // Build structured graph
-    const structuredNodes = entries
-      .filter((entry: NoteEntry) => (entry.note.relatedTo?.length ?? 0) > 0)
-      .slice(0, limit)
-      .map((entry: NoteEntry) => {
-        const edges = (entry.note.relatedTo ?? [])
-          .filter((rel) => visibleIds.has(rel.id))
-          .map((rel) => ({ toId: rel.id, type: rel.type }));
-        return {
-          id: entry.note.id,
-          title: entry.note.title,
-          edges: edges.length > 0 ? edges : [],
-        };
-      })
-      .filter((node: { edges: any[] }) => node.edges.length > 0);
-    
-    const structuredContent: MemoryGraphResult = {
-      action: "graph_shown",
-      project: toProjectRef(project),
-      nodes: structuredNodes,
-      limit,
-      truncated: structuredNodes.length < entries.filter(e => (e.note.relatedTo?.length ?? 0) > 0).length,
-    };
-    
-    return { content: [{ type: "text", text: textContent }], structuredContent };
-  }
-);
-
-// ── project_memory_summary ────────────────────────────────────────────────────
-server.registerTool(
-  "project_memory_summary",
-  {
-    title: "Project Memory Summary",
-    description:
-      "Generate a high-level summary of what is already known for a project.\n\n" +
-      "Use this when:\n" +
-      "- You want fast orientation before starting work\n" +
-      "- You need a synthesized overview instead of raw note listings\n\n" +
-      "Do not use this when:\n" +
-      "- You need exact note contents; use `get`\n" +
-      "- You need direct semantic matches for a query; use `recall`\n\n" +
-      "Returns:\n" +
-      "- A synthesized project-level summary based on stored memories\n" +
-      "- Bounded 1-hop relationship previews on orientation entry points (primaryEntry and suggestedNext)\n\n" +
-      "- Optional compact working-state recovery hints when relevant temporary notes exist\n\n" +
-      "Read-only.\n\n" +
-      "Typical next step:\n" +
-      "- Use `recall` or `list` to drill down into specific areas.",
-    annotations: {
-      readOnlyHint: true,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      cwd: z.string().describe("Absolute project working directory. Pass this whenever the task is project-related so routing, search boosting, policy, and vault selection work correctly."),
-      maxPerTheme: z.number().int().min(1).max(5).optional().default(3),
-      recentLimit: z.number().int().min(1).max(10).optional().default(5),
-      anchorLimit: z.number().int().min(1).max(10).optional().default(5),
-      includeRelatedGlobal: z.boolean().optional().default(false),
-      relatedGlobalLimit: z.number().int().min(1).max(5).optional().default(3),
-    }),
-    outputSchema: ProjectSummaryResultSchema,
-  },
-  async ({ cwd, maxPerTheme, recentLimit, anchorLimit, includeRelatedGlobal, relatedGlobalLimit }) => {
-    const t0Summary = performance.now();
-    await ensureBranchSynced(cwd);
-
-    // Pre-resolve project so we can pass its id to collectVisibleNotes for session caching
-    const preProject = await resolveProject(cwd);
-    const { project, entries } = await collectVisibleNotes(cwd, "all", undefined, "any", preProject?.id);
-    if (!project) {
-      return projectNotFoundResponse(cwd);
-    }
-
-    // Separate project-scoped notes (for themes/anchors) from global notes
-    const projectEntries = entries.filter(e =>
-      e.note.project === project.id || e.vault.isProject
-    );
-
-    // Empty-project case: no project-scoped notes exist
-    if (projectEntries.length === 0) {
-      const structuredContent: ProjectSummaryResult = {
-        action: "project_summary_shown",
-        project: { id: project.id, name: project.name },
-        notes: { total: 0, projectVault: 0, mainVault: 0, privateProject: 0 },
-        themes: {},
-        recent: [],
-        anchors: [],
-        orientation: {
-          primaryEntry: { id: "", title: "No notes", rationale: "Empty project vault" },
-          suggestedNext: [],
-        },
-      };
-      return { content: [{ type: "text", text: `No memories found for project ${project.name}.` }], structuredContent };
-    }
-
-    const policyLine = await formatProjectPolicyLine(project.id);
-
-    const projectNoteIds = new Set(projectEntries.map(e => e.note.id));
-
-    // Compute promoted themes from keywords (graduation system)
-    const graduationResult = computeThemesWithGraduation(projectEntries.map(e => e.note));
-    const promotedThemes = new Set(graduationResult.promotedThemes);
-    const themeCache = graduationResult.themeAssignments;
-
-    const inboundReferences = new Map<string, number>();
-    const linkedByPermanentNotes = new Map<string, number>();
-    for (const entry of projectEntries) {
-      for (const rel of entry.note.relatedTo ?? []) {
-        if (!projectNoteIds.has(rel.id)) {
-          continue;
-        }
-
-        inboundReferences.set(rel.id, (inboundReferences.get(rel.id) ?? 0) + 1);
-        if (entry.note.lifecycle === "permanent") {
-          linkedByPermanentNotes.set(rel.id, (linkedByPermanentNotes.get(rel.id) ?? 0) + 1);
-        }
-      }
-    }
-
-    const effectiveMetadataById = new Map(
-      projectEntries.map((entry) => {
-        const inbound = inboundReferences.get(entry.note.id) ?? 0;
-        const visibleOutbound = (entry.note.relatedTo ?? []).filter((rel) => projectNoteIds.has(rel.id)).length;
-        const metadata = getEffectiveMetadata(entry.note, {
-          inboundReferences: inbound,
-          linkedByPermanentNotes: linkedByPermanentNotes.get(entry.note.id) ?? 0,
-          anchorCandidate: entry.note.lifecycle === "permanent" && (visibleOutbound > 0 || inbound > 0),
-        });
-        return [entry.note.id, {
-          metadata,
-          inbound,
-          visibleOutbound,
-        }] as const;
-      })
-    );
-
-    // Categorize by theme with graduation (project-scoped only)
-    const themed = new Map<string, NoteEntry[]>();
-    for (const entry of projectEntries) {
-      const theme = classifyThemeWithGraduation(entry.note, promotedThemes);
-      const bucket = themed.get(theme) ?? [];
-      bucket.push(entry);
-      themed.set(theme, bucket);
-    }
-
-    // Theme order: fixed themes first, then promoted themes alphabetically, then "other"
-    const fixedThemes = ["overview", "decisions", "tooling", "bugs", "architecture", "quality"];
-    const dynamicThemes = graduationResult.promotedThemes.filter(t => !fixedThemes.includes(t));
-    const themeOrder = [...fixedThemes, ...dynamicThemes.sort(), "other"];
-
-    // Collapse thin dynamic-theme buckets (< 2 notes) into "other" to reduce noise.
-    // Fixed themes are kept even when small; only keyword-graduated themes are collapsed.
-    const fixedThemeSet = new Set(fixedThemes);
-    for (const [theme, bucket] of Array.from(themed.entries())) {
-      if (!fixedThemeSet.has(theme) && theme !== "other" && bucket.length < 2) {
-        const otherBucket = themed.get("other") ?? [];
-        otherBucket.push(...bucket);
-        themed.set("other", otherBucket);
-        themed.delete(theme);
-      }
-    }
-
-    // Calculate notes distribution (project-scoped only)
-    const projectVaultCount = projectEntries.filter(e => e.vault.isProject).length;
-    const mainVaultProjectEntries = projectEntries.filter(e => !e.vault.isProject);
-    const mainVaultCount = mainVaultProjectEntries.length;
-    const totalProjectNotes = projectEntries.length;
-
-    // Build output sections
-    const sections: string[] = [];
-    sections.push(`Project summary: **${project.name}**`);
-    sections.push(`- id: \`${project.id}\``);
-    sections.push(`- ${policyLine.replace(/^Policy:\s*/, "policy: ")}`);
-    sections.push(`- memories: ${totalProjectNotes} (project-vault: ${projectVaultCount}, main-vault: ${mainVaultCount})`);
-    if (mainVaultProjectEntries.length > 0) {
-      sections.push(`- private project memories: ${mainVaultProjectEntries.length}`);
-    }
-
-    const themes: Record<string, ThemeSection> = {};
-    for (const theme of themeOrder) {
-      const bucket = themed.get(theme);
-      if (!bucket || bucket.length === 0) continue;
-      
-      // Sort by within-theme score
-      const sorted = [...bucket].sort((a, b) => 
-        withinThemeScore(b.note, effectiveMetadataById.get(b.note.id)?.metadata) - withinThemeScore(a.note, effectiveMetadataById.get(a.note.id)?.metadata)
-      );
-      const top = sorted.slice(0, maxPerTheme);
-      
-      sections.push(`\n${titleCaseTheme(theme)}:`);
-      sections.push(...top.map(e => `- ${e.note.title} (\`${e.note.id}\`)`));
-      
-      themes[theme] = {
-        count: bucket.length,
-        examples: top.map(e => ({
-          id: e.note.id,
-          title: e.note.title,
-          updatedAt: e.note.updatedAt,
-        })),
-      };
-    }
-
-    // Recent notes (project-scoped only)
-    const recent = [...projectEntries]
-      .sort((a, b) => b.note.updatedAt.localeCompare(a.note.updatedAt))
-      .slice(0, recentLimit);
-    
-    sections.push(`\nRecent activity (start here):`);
-    sections.push(...recent.map(e => `- ${e.note.updatedAt} — ${e.note.title}`));
-
-    const temporaryEntries = projectEntries
-      .filter((entry) => entry.note.lifecycle === "temporary")
-      .map((entry) => {
-        const metadata = effectiveMetadataById.get(entry.note.id)?.metadata;
-        const score = workingStateScore(entry.note, metadata);
-        const nextAction = extractNextAction(entry.note);
-        const relatedCount = entry.note.relatedTo?.length ?? 0;
-        const days = daysSinceUpdate(entry.note.updatedAt);
-        const rationaleParts = [`updated ${days < 1 ? "today" : `${Math.round(days)}d ago`}`];
-        if (relatedCount > 0) rationaleParts.push(`${relatedCount} linked note${relatedCount === 1 ? "" : "s"}`);
-        if (nextAction) rationaleParts.push("explicit next action");
-        if (metadata?.role === "plan" || metadata?.role === "context") rationaleParts.push(`${metadata.role} note`);
-
-        return {
-          entry,
-          score,
-          rationale: rationaleParts.join(", "),
-          preview: summarizePreview(entry.note.content, 120),
-          nextAction,
-        };
-      })
-      .filter((candidate) => candidate.score > -Infinity)
-      .sort((a, b) => b.score - a.score || b.entry.note.updatedAt.localeCompare(a.entry.note.updatedAt))
-      .slice(0, 3);
-
-    const workingState = temporaryEntries.length > 0
-      ? {
-          summary:
-            temporaryEntries.length === 1
-              ? `1 temporary note may help resume active work.`
-              : `${temporaryEntries.length} temporary notes may help resume active work.`,
-          recoveryHint: "Orient with project_memory_summary first, then inspect these temporary notes if you need to continue in-progress work.",
-          notes: temporaryEntries.map(({ entry, rationale, preview, nextAction }) => ({
-            id: entry.note.id,
-            title: entry.note.title,
-            updatedAt: entry.note.updatedAt,
-            rationale,
-            preview,
-            nextAction,
-          })),
-        }
-      : undefined;
-
-    // Anchor notes with diversity constraint (project-scoped only)
-    const scoredAnchorCandidates = projectEntries
-      .map(e => {
-        const baselineContext = effectiveMetadataById.get(e.note.id);
-        const metadata = baselineContext?.metadata;
-        return {
-          entry: e,
-          metadata,
-          score: anchorScore(e.note, themeCache, metadata),
-          theme: themeCache.get(e.note.id) ?? "other",
-          alwaysLoad: metadata?.alwaysLoad === true,
-          explicitOrientationRole:
-            metadata?.roleSource === "explicit" &&
-            (metadata.role === "summary" || metadata.role === "decision"),
-          hasVisibleGraphParticipation: (baselineContext?.visibleOutbound ?? 0) > 0 || (baselineContext?.inbound ?? 0) > 0,
-        };
-      })
-      .filter(candidate => candidate.score > -Infinity)
-      .filter(candidate => candidate.alwaysLoad || candidate.explicitOrientationRole || candidate.hasVisibleGraphParticipation)
-      .sort((a, b) => b.score - a.score || a.entry.note.title.localeCompare(b.entry.note.title));
-
-    // Enforce max 2 per theme for scored anchors
-    const anchorThemeCounts = new Map<string, number>();
-    const anchors: AnchorNote[] = [];
-    const anchorIds = new Set<string>();
-
-    // Add scored anchors with theme diversity constraint
-    for (const candidate of scoredAnchorCandidates) {
-      if (anchors.length >= 10) break;
-
-      const theme = candidate.theme;
-      const themeCount = anchorThemeCounts.get(theme) ?? 0;
-      if (themeCount >= 2) continue;
-
-      anchors.push({
-        id: candidate.entry.note.id,
-        title: candidate.entry.note.title,
-        centrality: candidate.entry.note.relatedTo?.length ?? 0,
-        connectionDiversity: computeConnectionDiversity(candidate.entry.note, themeCache),
-        updatedAt: candidate.entry.note.updatedAt,
-      });
-      anchorIds.add(candidate.entry.note.id);
-      anchorThemeCounts.set(theme, themeCount + 1);
-    }
-
-    if (anchors.length > 0) {
-      sections.push(`\nAnchors:`);
-      sections.push(...anchors.slice(0, 5).map(a => 
-        `- ${a.title} (\`${a.id}\`) — centrality: ${a.centrality}, diversity: ${a.connectionDiversity}`
-      ));
-    }
-
-    // Compute orientation after anchors are computed (for text output)
-    let relatedGlobal: ProjectSummaryResult["relatedGlobal"];
-    
-    if (includeRelatedGlobal) {
-      const anchorEmbeddings = await Promise.all(
-        anchors.slice(0, 5).map(async a => {
-          for (const vault of vaultManager.allKnownVaults()) {
-            const emb = await vault.storage.readEmbedding(memoryId(a.id));
-            if (emb) return { id: a.id, embedding: emb.embedding };
-          }
-          return null;
-        })
-      );
-      
-      const validAnchors = anchorEmbeddings.filter((e): e is NonNullable<typeof e> => e !== null);
-      
-      if (validAnchors.length > 0) {
-        // Get global notes (not project-scoped)
-        const globalEntries = entries.filter(e => !e.note.project);
-        const globalCandidates: Array<{ id: string; title: string; similarity: number; preview: string }> = [];
-        
-        for (const entry of globalEntries) {
-          const emb = await entry.vault.storage.readEmbedding(entry.note.id);
-          if (!emb) continue;
-          
-          // Find max similarity to any anchor
-          let maxSim = 0;
-          for (const anchor of validAnchors) {
-            const sim = cosineSimilarity(anchor.embedding, emb.embedding);
-            if (sim > maxSim) maxSim = sim;
-          }
-          
-          if (maxSim > 0.4) {
-            const projection = await entry.vault.storage.readProjection(entry.note.id);
-            const preview = projection?.summary
-              ? projection.summary.slice(0, 100)
-              : summarizePreview(entry.note.content, 100);
-            globalCandidates.push({
-              id: entry.note.id,
-              title: entry.note.title,
-              similarity: maxSim,
-              preview,
-            });
-          }
-        }
-        
-        globalCandidates.sort((a, b) => b.similarity - a.similarity);
-        
-        if (globalCandidates.length > 0) {
-          relatedGlobal = {
-            notes: globalCandidates.slice(0, relatedGlobalLimit),
-            computedAt: new Date().toISOString(),
-          };
-          
-          sections.push(`\nRelated Global:`);
-          sections.push(...relatedGlobal.notes.map(n => 
-            `- ${n.title} (\`${n.id}\`) — similarity: ${n.similarity.toFixed(2)}`
-          ));
-        }
-      }
-    }
-
-    // Compute orientation layer for actionable guidance
-    const primaryAnchor = anchors[0];
-
-    // Build noteId -> vault lookup for provenance enrichment
-    const noteVaultMap = new Map<string, Vault>();
-    for (const entry of projectEntries) {
-      noteVaultMap.set(entry.note.id, entry.vault);
-    }
-
-    // Helper to enrich an anchor with provenance and confidence
-    const enrichOrientationNote = async (anchor: AnchorNote) => {
-      const vault = noteVaultMap.get(anchor.id);
-      if (!vault) return {};
-      const filePath = `${vault.notesRelDir}/${anchor.id}.md`;
-      const provenance = await getNoteProvenance(vault.git, filePath);
-      const confidence = computeConfidence("permanent", anchor.updatedAt, anchor.centrality);
-      return { provenance, confidence };
-    };
-
-    // Helper to enrich an anchor with relationships (1-hop expansion)
-    const enrichOrientationNoteWithRelationships = async (anchor: AnchorNote) => {
-      const vault = noteVaultMap.get(anchor.id);
-      if (!vault) return {};
-      const note = await vault.storage.readNote(memoryId(anchor.id));
-      if (!note) return {};
-      const relationships = await getRelationshipPreview(
-        note,
-        vaultManager.allKnownVaults(),
-        { activeProjectId: project.id, limit: 3 }
-      );
-      return { relationships };
-    };
-
-    const primaryEnriched = primaryAnchor ? await enrichOrientationNote(primaryAnchor) : {};
-    const primaryRelationships = primaryAnchor ? await enrichOrientationNoteWithRelationships(primaryAnchor) : {};
-
-    // Select theme-diverse suggestedNext: avoid repeating the primary anchor's theme.
-    // Backfills without constraint if not enough theme-distinct candidates exist.
-    const primaryTheme = primaryAnchor ? (themeCache.get(primaryAnchor.id) ?? "other") : "other";
-    const usedSuggestedThemes = new Set([primaryTheme]);
-    const suggestedCandidates: typeof anchors = [];
-    for (const anchor of anchors.slice(1)) {
-      if (suggestedCandidates.length >= 3) break;
-      const anchorTheme = themeCache.get(anchor.id) ?? "other";
-      if (!usedSuggestedThemes.has(anchorTheme)) {
-        suggestedCandidates.push(anchor);
-        usedSuggestedThemes.add(anchorTheme);
-      }
-    }
-    for (const anchor of anchors.slice(1)) {
-      if (suggestedCandidates.length >= 3) break;
-      if (!suggestedCandidates.includes(anchor)) {
-        suggestedCandidates.push(anchor);
-      }
-    }
-
-    const suggestedEnriched = await Promise.all(suggestedCandidates.map(enrichOrientationNote));
-    const suggestedRelationships = await Promise.all(suggestedCandidates.map(enrichOrientationNoteWithRelationships));
-
-    const recentPermanent = recent.find((entry) => entry.note.lifecycle === "permanent");
-    const fallbackEntry = recentPermanent ?? recent[0];
-    const permanentOverrideUsed = Boolean(recentPermanent && recent[0] && recentPermanent.note.id !== recent[0].note.id);
-
-    // Enrich fallback primaryEntry when no anchors exist
-    let fallbackEnriched: { provenance?: { lastUpdatedAt: string; lastCommitHash: string; lastCommitMessage: string; recentlyChanged: boolean }; confidence?: Confidence } = {};
-    let fallbackRelationships: { relationships?: RelationshipPreview } = {};
-    if (!primaryAnchor && fallbackEntry) {
-      const fallbackNote = fallbackEntry.note;
-      const vault = noteVaultMap.get(fallbackNote.id);
-      if (vault) {
-        const filePath = `${vault.notesRelDir}/${fallbackNote.id}.md`;
-        const provenance = await getNoteProvenance(vault.git, filePath);
-        const confidence = computeConfidence(fallbackNote.lifecycle, fallbackNote.updatedAt, 0);
-        fallbackEnriched = { provenance, confidence };
-      }
-      const preview = await getRelationshipPreview(
-        fallbackNote,
-        vaultManager.allKnownVaults(),
-        { activeProjectId: project.id, limit: 3 }
-      );
-      if (preview) fallbackRelationships = { relationships: preview };
-    }
-
-    const orientation: ProjectSummaryResult["orientation"] = {
-      primaryEntry: primaryAnchor
-        ? {
-            id: primaryAnchor.id,
-            title: primaryAnchor.title,
-            rationale: `Centrality ${primaryAnchor.centrality}, connects ${primaryAnchor.connectionDiversity} themes`,
-            ...primaryEnriched,
-            ...primaryRelationships,
-          }
-        : {
-            id: fallbackEntry?.note.id ?? projectEntries[0]?.note.id ?? "",
-            title: fallbackEntry?.note.title ?? projectEntries[0]?.note.title ?? "No notes",
-            rationale: permanentOverrideUsed
-              ? "Most recent permanent note — no high-centrality anchors found"
-              : fallbackEntry
-              ? "Most recent note — no high-centrality anchors found"
-              : "Only note available",
-            ...fallbackEnriched,
-            ...fallbackRelationships,
-          },
-      suggestedNext: suggestedCandidates.map((a, i) => ({
-        id: a.id,
-        title: a.title,
-        rationale: `Centrality ${a.centrality}, connects ${a.connectionDiversity} themes`,
-        ...suggestedEnriched[i],
-        ...suggestedRelationships[i],
-      })),
-    };
-
-    // Warning for taxonomy dilution
-    const otherBucket = themed.get("other");
-    const otherCount = otherBucket?.length ?? 0;
-    const otherRatio = projectEntries.length > 0 ? otherCount / projectEntries.length : 0;
-    if (otherRatio > 0.3) {
-      orientation.warnings = [
-        `${Math.round(otherRatio * 100)}% of notes in "other" bucket — consider improving thematic classification`,
-      ];
-    }
-
-    // Orientation text output
-    sections.push(`\nOrientation:`);
-    sections.push(`Start with: ${orientation.primaryEntry.title} (\`${orientation.primaryEntry.id}\`)`);
-    sections.push(`  Rationale: ${orientation.primaryEntry.rationale}`);
-    if (orientation.primaryEntry.confidence) {
-      sections.push(`  Confidence: ${orientation.primaryEntry.confidence}`);
-    }
-    if (orientation.primaryEntry.relationships) {
-      sections.push(`  ${formatRelationshipPreview(orientation.primaryEntry.relationships)}`);
-    }
-    if (orientation.suggestedNext.length > 0) {
-      sections.push(`Then check:`);
-      for (const next of orientation.suggestedNext) {
-        sections.push(`  - ${next.title} (\`${next.id}\`) — ${next.rationale}${next.confidence ? ` [${next.confidence}]` : ""}`);
-        if (next.relationships) {
-          sections.push(`    ${formatRelationshipPreview(next.relationships)}`);
-        }
-      }
-    }
-    if (orientation.warnings && orientation.warnings.length > 0) {
-      sections.push(`Warnings:`);
-      for (const w of orientation.warnings) {
-        sections.push(`  - ${w}`);
-      }
-    }
-
-    if (workingState) {
-      sections.push(`\nWorking state:`);
-      sections.push(workingState.summary);
-      sections.push(`Recovery hint: ${workingState.recoveryHint}`);
-      for (const note of workingState.notes) {
-        sections.push(`- ${note.title} (\`${note.id}\`) — ${note.rationale}`);
-        sections.push(`  Preview: ${note.preview}`);
-        if (note.nextAction) {
-          sections.push(`  Next action: ${note.nextAction}`);
-        }
-      }
-    }
-
-    // Related global notes (optional, anchor-based similarity)
-
-    const structuredContent: ProjectSummaryResult = {
-      action: "project_summary_shown",
-      project: { id: project.id, name: project.name },
-      notes: {
-        total: totalProjectNotes,
-        projectVault: projectVaultCount,
-        mainVault: mainVaultCount,
-        privateProject: mainVaultProjectEntries.length,
-      },
-      themes,
-      recent: recent.map(e => ({
-        id: e.note.id,
-        title: e.note.title,
-        updatedAt: e.note.updatedAt,
-        theme: classifyThemeWithGraduation(e.note, promotedThemes),
-      })),
-      anchors,
-      orientation,
-      workingState,
-      relatedGlobal,
-    };
-
-    console.error(`[summary:timing] ${(performance.now() - t0Summary).toFixed(1)}ms`);
-    return { content: [{ type: "text", text: sections.join("\n") }], structuredContent };
-  }
-);
-
-// ── sync ──────────────────────────────────────────────────────────────────────
-server.registerTool(
-  "sync",
-  {
-    title: "Sync",
-    description:
-      "Sync mnemonic vaults with their git remotes and repair local embedding coverage as needed.\n\n" +
-      "Use this when:\n" +
-      "- You want the local vault state aligned with remote changes\n" +
-      "- You suspect another machine or collaborator updated the vaults\n\n" +
-      "Do not use this when:\n" +
-      "- You only need to inspect or edit a single memory\n\n" +
-      "Returns:\n" +
-      "- Per-vault pull/push results, deletions, additions, and embedding rebuild info\n\n" +
-      "Side effects: performs git sync and may rebuild local embeddings.\n\n" +
-      "Typical next step:\n" +
-      "- Use `recent_memories`, `list`, or `recall` to inspect newly synced changes.",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: true,
-    },
-    inputSchema: z.object({
-      cwd: projectParam,
-      force: z.boolean().optional().default(false).describe("Rebuild all embeddings even if the current model already generated them"),
-    }),
-    outputSchema: SyncResultSchema,
-  },
-  async ({ cwd, force }) => {
-    const lines: string[] = [];
-    const vaultResults: Array<StructuredSyncResult["vaults"][number]> = [];
-
-    // Always sync main vault
-    const mainVaultPath = vaultManager.main.storage.vaultPath;
-    const mainResult = await vaultManager.main.git.sync();
-    lines.push(...formatSyncResult(mainResult, "main vault", mainVaultPath));
-    let mainEmbedded = 0;
-    let mainFailed: string[] = [];
-    const mainBackfill = await backfillEmbeddingsAfterSync(vaultManager.main.storage, "main vault", lines, force);
-    mainEmbedded = mainBackfill.embedded;
-    mainFailed = mainBackfill.failed;
-    if (mainResult.deletedNoteIds.length > 0) {
-      await removeStaleEmbeddings(vaultManager.main.storage, mainResult.deletedNoteIds);
-    }
-    vaultResults.push({
-      vault: "main",
-      hasRemote: mainResult.hasRemote,
-      pulled: mainResult.pulledNoteIds.length,
-      deleted: mainResult.deletedNoteIds.length,
-      pushed: mainResult.pushedCommits,
-      embedded: mainEmbedded,
-      failed: mainFailed,
-      gitError: mainResult.gitError,
-    });
-
-    // Optionally sync project vault
-    if (cwd) {
-      const projectVault = await vaultManager.getProjectVaultIfExists(cwd);
-      if (projectVault) {
-        const projectVaultPath = projectVault.storage.vaultPath;
-        const projectResult = await projectVault.git.sync();
-        lines.push(...formatSyncResult(projectResult, "project vault", projectVaultPath));
-        let projEmbedded = 0;
-        let projFailed: string[] = [];
-        const projectBackfill = await backfillEmbeddingsAfterSync(projectVault.storage, "project vault", lines, force);
-        projEmbedded = projectBackfill.embedded;
-        projFailed = projectBackfill.failed;
-        if (projectResult.deletedNoteIds.length > 0) {
-          await removeStaleEmbeddings(projectVault.storage, projectResult.deletedNoteIds);
-        }
-        vaultResults.push({
-          vault: "project",
-          hasRemote: projectResult.hasRemote,
-          pulled: projectResult.pulledNoteIds.length,
-          deleted: projectResult.deletedNoteIds.length,
-          pushed: projectResult.pushedCommits,
-          embedded: projEmbedded,
-          failed: projFailed,
-          gitError: projectResult.gitError,
-        });
-      } else {
-        lines.push("project vault: no .mnemonic/ found — skipped.");
-      }
-    }
-
-    const structuredContent: StructuredSyncResult = {
-      action: "synced",
-      vaults: vaultResults,
-    };
-
-    // Vault contents may have changed via pull — discard session cache
-    invalidateActiveProjectCache();
-    return { content: [{ type: "text", text: lines.join("\n") }], structuredContent };
-  }
-);
-
-// ── move_memory ───────────────────────────────────────────────────────────────
-server.registerTool(
-  "move_memory",
-  {
-    title: "Move Memory",
-    description:
-      "Use after `where_is_memory` or `get` confirms a memory is stored in the wrong place.\n\n" +
-      "Move a memory between main storage and project storage, optionally targeting a specific sub-vault folder.\n\n" +
-      "Use this when:\n" +
-      "- A note is stored in the wrong place\n" +
-      "- Project-specific knowledge should move between shared project storage and main storage\n" +
-      "- A note should live in a specific sub-vault such as `.mnemonic-lib`\n\n" +
-      "Do not use this when:\n" +
-      "- You only need to edit the note content; use `update`\n" +
-      "- You want to delete the note; use `forget`\n\n" +
-      "Returns:\n" +
-      "- The moved memory id, resulting storage label, project association, and persistence status\n\n" +
-      "Side effects: rewrites storage location, may adjust project association, git commits, and may push.\n\n" +
-      "Typical next step:\n" +
-      "- Use `where_is_memory` or `get` to verify the final state.\n" +
-      "- Use `relate` if the moved memory connects to existing notes in the new vault.",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      id: NoteIdSchema.describe("Exact memory id. Use an id returned by `recall`, `list`, `recent_memories`, or `where_is`."),
-      target: z.enum(["main-vault", "project-vault"]).describe("Destination: 'main-vault' for private/global storage, 'project-vault' for shared project storage"),
-      vaultFolder: z
-        .string()
-        .optional()
-        .describe(
-          "Optional target project vault folder name, such as `.mnemonic-lib`. " +
-          "Use this only when moving into a specific sub-vault instead of the primary project vault."
-        ),
-      cwd: projectParam,
-      allowProtectedBranch: z
-        .boolean()
-        .optional()
-        .describe(
-          "One-time override for protected branch checks. " +
-          "When true, move_memory can commit on a protected branch without changing project policy."
-        ),
-    }),
-    outputSchema: MoveResultSchema,
-  },
-  async ({ id, target, vaultFolder, cwd, allowProtectedBranch = false }) => {
-    await ensureBranchSynced(cwd);
-
-    const found = await vaultManager.findNote(id, cwd);
-    if (!found) {
-      return { content: [{ type: "text", text: `No memory found with id '${id}'` }], isError: true };
-    }
-
-    const currentStorage = storageLabel(found.vault);
-
-    let targetVault: Vault;
-    let targetProject: Awaited<ReturnType<typeof resolveProject>> | undefined;
-    if (target === "main-vault") {
-      targetVault = vaultManager.main;
-    } else {
-      if (!cwd) {
-        return {
-          content: [{
-            type: "text",
-            text: "Moving into a project vault requires `cwd` so mnemonic can resolve the destination project.",
-          }],
-          isError: true,
-        };
-      }
-
-      if (vaultFolder) {
-        // Target a specific submodule vault by folder name.
-        const subVault = await vaultManager.getVaultByFolder(cwd, vaultFolder);
-        if (!subVault) {
-          return {
-            content: [{ type: "text", text: `Vault folder '${vaultFolder}' not found under the git root for: ${cwd}` }],
-            isError: true,
-          };
-        }
-        targetVault = subVault;
-      } else {
-        // Default: primary project vault (.mnemonic).
-        const projectVault = await vaultManager.getOrCreateProjectVault(cwd);
-        if (!projectVault) {
-          return { content: [{ type: "text", text: `Could not resolve a project vault for: ${cwd}` }], isError: true };
-        }
-        targetVault = projectVault;
-      }
-
-      targetProject = await resolveProject(cwd);
-      if (!targetProject) {
-        return projectNotFoundResponse(cwd);
-      }
-    }
-
-    // Check if the note is already in the target vault.
-    if (found.vault.storage.vaultPath === targetVault.storage.vaultPath) {
-      const targetLabel = storageLabel(targetVault);
-      return { content: [{ type: "text", text: `Memory '${id}' is already stored in ${targetLabel}.` }], isError: true };
-    }
-
-    if (found.vault.isProject || targetVault.isProject) {
-      const resolvedProject = targetProject ?? await resolveProject(cwd);
-      const projectLabel = resolvedProject
-        ? `${resolvedProject.name} (${resolvedProject.id})`
-        : `${found.note.projectName ?? "project"} (${found.note.project ?? "unknown"})`;
-      const projectId = targetProject?.id ?? found.note.project;
-      const policy = projectId ? await configStore.getProjectPolicy(projectId) : undefined;
-      const protectedBranchCheck = await shouldBlockProtectedBranchCommit({
-        cwd,
-        writeScope: "project",
-        automaticCommit: true,
-        projectLabel,
-        policy,
-        allowProtectedBranch,
-        toolName: "move_memory",
-      });
-      if (protectedBranchCheck.blocked) {
-        return {
-          content: [{ type: "text", text: protectedBranchCheck.message ?? "Protected branch policy blocked this commit." }],
-          isError: true,
-        };
-      }
-    }
-
-    const targetLabel = storageLabel(targetVault);
-    const existing = await targetVault.storage.readNote(memoryId(id));
-    if (existing) {
-      return { content: [{ type: "text", text: `Cannot move '${id}' because a note with that id already exists in ${targetLabel}.` }], isError: true };
-    }
-
-    let noteToWrite = found.note;
-    let metadataRewritten = false;
-    if (target === "project-vault" && targetProject) {
-      const rewrittenProject = targetProject.id;
-      const rewrittenProjectName = targetProject.name;
-      metadataRewritten = noteToWrite.project !== rewrittenProject || noteToWrite.projectName !== rewrittenProjectName;
-      noteToWrite = {
-        ...noteToWrite,
-        project: rewrittenProject,
-        projectName: rewrittenProjectName,
-        updatedAt: isoDateString(new Date().toISOString()),
-      };
-    }
-
-    const moveResult = await moveNoteBetweenVaults(found, targetVault, noteToWrite, cwd);
-    const movedNote = moveResult.note;
-    const associationValue = movedNote.projectName && movedNote.project
-      ? `${movedNote.projectName} (${movedNote.project})`
-      : movedNote.projectName ?? movedNote.project ?? "global";
-    
-    const structuredContent: MoveResult = {
-      action: "moved",
-      id,
-      fromVault: currentStorage,
-      toVault: targetLabel,
-      projectAssociation: associationValue,
-      title: movedNote.title,
-      metadataRewritten,
-      persistence: moveResult.persistence,
-    };
-
-    const associationText = metadataRewritten
-      ? `Project association is now ${associationValue}.`
-      : `Project association remains ${associationValue}.`;
-
-    invalidateActiveProjectCache();
-    return {
-      content: [{
-        type: "text",
-        text: `Moved '${id}' from ${currentStorage} to ${targetLabel}. ${associationText}\n${formatPersistenceSummary(moveResult.persistence)}`,
-      }],
-      structuredContent,
-    };
-  }
-);
-
-// ── relate ────────────────────────────────────────────────────────────────────
-server.registerTool(
-  "relate",
-  {
-    title: "Relate Memories",
-    description:
-      "Use after you have identified the exact memories to connect.\n\n" +
-      "Create a typed bidirectional relationship between two memories.\n\n" +
-      "Use this when:\n" +
-      "- A newly stored or updated note meaningfully connects to another note\n" +
-      "- One note explains, exemplifies, supersedes, or closely relates to another\n\n" +
-      "Do not use this when:\n" +
-      "- The connection is weak or speculative\n" +
-      "- You need to remove a relationship rather than add one\n\n" +
-      "Returns:\n" +
-      "- Both memory ids and the created relationship type\n\n" +
-      "Side effects: modifies both notes, git commits per affected vault, and may push.\n\n" +
-      "Typical next step:\n" +
-      "- Use `get` on both notes to verify the relationship context reads well.",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: false,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      fromId: z.string().describe("Source memory id"),
-      toId: z.string().describe("Target memory id"),
-      type: z.enum(RELATIONSHIP_TYPES).default("related-to").describe(
-        "Relationship type: 'related-to' (same topic), 'explains' (clarifies why), 'example-of' (instance of pattern), 'supersedes' (replaces), 'derives-from' (derived artifact), 'follows' (sequence order)"
-      ),
-      bidirectional: z.boolean().optional().default(true).describe("Add relationship in both directions (default: true)"),
-      cwd: projectParam,
-    }),
-    outputSchema: RelateResultSchema,
-  },
-  async ({ fromId, toId, type, bidirectional, cwd }) => {
-    await ensureBranchSynced(cwd);
-
-    const [foundFrom, foundTo] = await Promise.all([
-      vaultManager.findNote(fromId, cwd),
-      vaultManager.findNote(toId, cwd),
-    ]);
-    if (!foundFrom) return { content: [{ type: "text", text: `No memory found with id '${fromId}'` }], isError: true };
-    if (!foundTo) return { content: [{ type: "text", text: `No memory found with id '${toId}'` }], isError: true };
-
-    const { note: fromNote, vault: fromVault } = foundFrom;
-    const { note: toNote, vault: toVault } = foundTo;
-    const now = isoDateString(new Date().toISOString());
-
-    // Group changes by vault so notes in the same vault share one commit
-    const vaultChanges = new Map<Vault, string[]>();
-
-    const fromRels = fromNote.relatedTo ?? [];
-    const fromRelExists = fromRels.some((r) => r.id === toId);
-    if (!fromRelExists) {
-      await fromVault.storage.writeNote({ ...fromNote, relatedTo: [...fromRels, { id: memoryId(toId), type }], updatedAt: now });
-      const files = vaultChanges.get(fromVault) ?? [];
-      files.push(vaultManager.noteRelPath(fromVault, fromId));
-      vaultChanges.set(fromVault, files);
-    }
-
-    const toRels = toNote.relatedTo ?? [];
-    const toRelExists = toRels.some((r) => r.id === fromId);
-    if (bidirectional && !toRelExists) {
-      await toVault.storage.writeNote({ ...toNote, relatedTo: [...toRels, { id: memoryId(fromId), type }], updatedAt: now });
-      const files = vaultChanges.get(toVault) ?? [];
-      files.push(vaultManager.noteRelPath(toVault, toId));
-      vaultChanges.set(toVault, files);
-    }
-
-    // Check for uncommitted changes from a previous failed attempt
-    if (vaultChanges.size === 0) {
-      // If relationships exist in note content, check git status for pending changes
-      const allVaults = new Set([fromVault, ...(bidirectional && toVault !== fromVault ? [toVault] : [])]);
-      
-      for (const vault of allVaults) {
-        const noteIds = vault === fromVault
-          ? (bidirectional && vault === toVault ? [fromId, toId] : [fromId])
-          : [toId];
-        const pendingFiles = await vaultManager.getPendingNoteFiles(vault, noteIds);
-        
-        if (pendingFiles.length > 0) {
-          // Commit the pending changes from previous failed attempt
-          const commitBody = formatCommitBody({
-            noteId: fromId,
-            noteTitle: fromNote.title,
-            projectName: fromNote.projectName,
-            relationship: { fromId, toId, type },
-          });
-          const commitMessage = `relate: ${fromNote.title} ↔ ${toNote.title}`;
-          const commitStatus = await vault.git.commitWithStatus(commitMessage, pendingFiles, commitBody);
-          
-          if (commitStatus.status === "committed") {
-            await pushAfterMutation(vault);
-          }
-          
-          const retry = buildMutationRetryContract({
-            commit: commitStatus,
-            commitMessage,
-            commitBody,
-            files: pendingFiles,
-            cwd,
-            vault,
-            mutationApplied: true,
-            preferredRecovery: "rerun-tool-call-serial",
-          });
-          
-          const structuredContent: RelateResult = {
-            action: "related",
-            fromId,
-            toId,
-            type,
-            bidirectional,
-            notesModified: pendingFiles.map((f: string) => path.basename(f, '.md')),
-            retry,
-          };
-          
-          const retrySummary = formatRetrySummary(retry);
-          return {
-            content: [{
-              type: "text",
-              text: `Reconciled pending commit for relationship \`${fromId}\` ${bidirectional ? "↔" : "→"} \`${toId}\` (${type})${retrySummary ? `\n${retrySummary}` : ""}`,
-            }],
-            structuredContent,
-          };
-        }
-      }
-      
-      return { content: [{ type: "text", text: `Relationship already exists between '${fromId}' and '${toId}'` }], isError: true };
-    }
-
-    const modifiedNoteIds: string[] = [];
-    let retry: MutationRetryContract | undefined;
-    for (const [vault, files] of vaultChanges) {
-      const isFromVault = vault === fromVault;
-      const thisNote = isFromVault ? fromNote : toNote;
-      const otherNote = isFromVault ? toNote : fromNote;
-      const commitBody = formatCommitBody({
-        noteId: thisNote.id,
-        noteTitle: thisNote.title,
-        projectName: thisNote.projectName,
-        relationship: {
-          fromId: thisNote.id,
-          toId: otherNote.id,
-          type,
-        },
-      });
-      const commitMessage = `relate: ${fromNote.title} ↔ ${toNote.title}`;
-      const commitStatus = await vault.git.commitWithStatus(commitMessage, files, commitBody);
-      if (!retry) {
-        retry = buildMutationRetryContract({
-          commit: commitStatus,
-          commitMessage,
-          commitBody,
-          files,
-          cwd,
-          vault,
-          mutationApplied: true,
-          preferredRecovery: "rerun-tool-call-serial",
-        });
-      }
-      if (commitStatus.status === "committed") {
-        await pushAfterMutation(vault);
-      }
-      modifiedNoteIds.push(...files.map(f => path.basename(f, '.md')));
-    }
-
-    const dirStr = bidirectional ? "↔" : "→";
-    const structuredContent: RelateResult = {
-      action: "related",
-      fromId,
-      toId,
-      type,
-      bidirectional,
-      notesModified: modifiedNoteIds,
-      retry,
-    };
-    
-    const retrySummary = formatRetrySummary(retry);
-    invalidateActiveProjectCache();
-    return {
-      content: [{
-        type: "text",
-        text: `Linked \`${fromId}\` ${dirStr} \`${toId}\` (${type})${retrySummary ? `\n${retrySummary}` : ""}`,
-      }],
-      structuredContent,
-    };
-  }
-);
-
-// ── unrelate ──────────────────────────────────────────────────────────────────
-server.registerTool(
-  "unrelate",
-  {
-    title: "Remove Relationship",
-    description:
-      "Remove a relationship between two memories.\n\n" +
-      "Use this when:\n" +
-      "- A previously created relationship is no longer accurate\n" +
-      "- Two notes should remain independent\n\n" +
-      "Do not use this when:\n" +
-      "- You are adding a new connection; use `relate`\n\n" +
-      "Returns:\n" +
-      "- Both memory ids and the removed relationship details\n\n" +
-      "Side effects: modifies both notes, git commits per affected vault, and may push.\n\n" +
-      "Typical next step:\n" +
-      "- Use `get` to verify both notes now stand on their own.",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: true,
-      idempotentHint: true,
-      openWorldHint: false,
-    },
-    inputSchema: z.object({
-      fromId: z.string().describe("Source memory id"),
-      toId: z.string().describe("Target memory id"),
-      bidirectional: z.boolean().optional().default(true).describe("Remove relationship in both directions (default: true)"),
-      cwd: projectParam,
-    }),
-    outputSchema: RelateResultSchema,
-  },
-  async ({ fromId, toId, bidirectional, cwd }) => {
-    await ensureBranchSynced(cwd);
-
-    const [foundFrom, foundTo] = await Promise.all([
-      vaultManager.findNote(fromId, cwd),
-      vaultManager.findNote(toId, cwd),
-    ]);
-
-    const now = isoDateString(new Date().toISOString());
-    const vaultChanges = new Map<Vault, string[]>();
-
-    if (foundFrom) {
-      const { note: fromNote, vault: fromVault } = foundFrom;
-      const filtered = (fromNote.relatedTo ?? []).filter((r) => r.id !== toId);
-      const fromRelExisted = (fromNote.relatedTo?.length ?? 0) > filtered.length;
-      if (fromRelExisted) {
-        await fromVault.storage.writeNote({ ...fromNote, relatedTo: filtered, updatedAt: now });
-        const files = vaultChanges.get(fromVault) ?? [];
-        files.push(vaultManager.noteRelPath(fromVault, fromId));
-        vaultChanges.set(fromVault, files);
-      }
-    }
-
-    if (bidirectional && foundTo) {
-      const { note: toNote, vault: toVault } = foundTo;
-      const filtered = (toNote.relatedTo ?? []).filter((r) => r.id !== fromId);
-      const toRelExisted = (toNote.relatedTo?.length ?? 0) > filtered.length;
-      if (toRelExisted) {
-        await toVault.storage.writeNote({ ...toNote, relatedTo: filtered, updatedAt: now });
-        const files = vaultChanges.get(toVault) ?? [];
-        files.push(vaultManager.noteRelPath(toVault, toId));
-        vaultChanges.set(toVault, files);
-      }
-    }
-
-    // Check for uncommitted changes from a previous failed attempt
-    if (vaultChanges.size === 0) {
-      // If no relationships were found in note content, check git status for pending changes
-      const allVaults = new Set<Vault>();
-      if (foundFrom) allVaults.add(foundFrom.vault);
-      if (bidirectional && foundTo) allVaults.add(foundTo.vault);
-      
-      for (const vault of allVaults) {
-        const noteIds: string[] = [];
-        if (foundFrom && foundFrom.vault === vault) noteIds.push(fromId);
-        if (bidirectional && foundTo && foundTo.vault === vault) noteIds.push(toId);
-        
-        const pendingFiles = await vaultManager.getPendingNoteFiles(vault, noteIds);
-        
-        if (pendingFiles.length > 0) {
-          // Commit the pending changes from previous failed attempt
-          const found = foundFrom?.vault === vault ? foundFrom : foundTo;
-          const commitBody = found
-            ? formatCommitBody({
-                noteId: found.note.id,
-                noteTitle: found.note.title,
-                projectName: found.note.projectName,
-              })
-            : undefined;
-          const commitMessage = `unrelate: ${fromId} ↔ ${toId}`;
-          const commitStatus = await vault.git.commitWithStatus(commitMessage, pendingFiles, commitBody);
-          
-          if (commitStatus.status === "committed") {
-            await pushAfterMutation(vault);
-          }
-          
-          const retry = buildMutationRetryContract({
-            commit: commitStatus,
-            commitMessage,
-            commitBody,
-            files: pendingFiles,
-            cwd,
-            vault,
-            mutationApplied: true,
-            preferredRecovery: "rerun-tool-call-serial",
-          });
-          
-          const structuredContent: RelateResult = {
-            action: "unrelated",
-            fromId,
-            toId,
-            type: "related-to",
-            bidirectional,
-            notesModified: pendingFiles.map((f: string) => path.basename(f, '.md')),
-            retry,
-          };
-          
-          const retrySummary = formatRetrySummary(retry);
-          return {
-            content: [{
-              type: "text",
-              text: `Reconciled pending commit for relationship removal between \`${fromId}\` and \`${toId}\`${retrySummary ? `\n${retrySummary}` : ""}`,
-            }],
-            structuredContent,
-          };
-        }
-      }
-      
-      return { content: [{ type: "text", text: `No relationship found between '${fromId}' and '${toId}'` }], isError: true };
-    }
-
-    let retry: MutationRetryContract | undefined;
-    for (const [vault, files] of vaultChanges) {
-      const found = foundFrom?.vault === vault ? foundFrom : foundTo;
-      const commitBody = found
-        ? formatCommitBody({
-            noteId: found.note.id,
-            noteTitle: found.note.title,
-            projectName: found.note.projectName,
-          })
-        : undefined;
-      const commitMessage = `unrelate: ${fromId} ↔ ${toId}`;
-      const commitStatus = await vault.git.commitWithStatus(commitMessage, files, commitBody);
-      if (!retry) {
-        retry = buildMutationRetryContract({
-          commit: commitStatus,
-          commitMessage,
-          commitBody,
-          files,
-          cwd,
-          vault,
-          mutationApplied: true,
-          preferredRecovery: "rerun-tool-call-serial",
-        });
-      }
-      if (commitStatus.status === "committed") {
-        await pushAfterMutation(vault);
-      }
-    }
-
-    const modifiedNoteIds: string[] = [];
-    for (const [vault, files] of vaultChanges) {
-      modifiedNoteIds.push(...files.map(f => path.basename(f, '.md')));
-    }
-    
-    const structuredContent: RelateResult = {
-      action: "unrelated",
-      fromId,
-      toId,
-      type: "related-to", // not tracked for unrelate
-      bidirectional,
-      notesModified: modifiedNoteIds,
-      retry,
-    };
-    
-    const retrySummary = formatRetrySummary(retry);
-    invalidateActiveProjectCache();
-    return {
-      content: [{
-        type: "text",
-        text: `Removed relationship between \`${fromId}\` and \`${toId}\`${retrySummary ? `\n${retrySummary}` : ""}`,
-      }],
-      structuredContent,
-    };
-  }
-);
-
-// ── consolidate ───────────────────────────────────────────────────────────────
-server.registerTool(
-  "consolidate",
-  {
-    title: "Consolidate Memories",
-    description:
-      "Use after `recall`, `list`, or `memory_graph` shows overlap that should be merged or cleaned up.\n\n" +
-      "Merge overlapping memories into a cleaner canonical note and retire the sources.\n\n" +
-      "Use this when:\n" +
-      "- Multiple notes cover the same decision, fix, or concept\n" +
-      "- One memory supersedes several older fragments\n\n" +
-      "Do not use this when:\n" +
-      "- You only need a small edit to one note; use `update`\n" +
-      "- You only want to connect notes; use `relate`\n\n" +
-      "Returns:\n" +
-      "- The canonical memory, source ids, resulting relationships, and persistence status\n\n" +
-      "Side effects: creates or updates the canonical note, modifies or removes source notes according to mode, git commits, and may push.\n\n" +
-      "Typical next step:\n" +
-      "- Use `get` to inspect the canonical note and `recall` to confirm duplication is reduced.\n" +
-      "- Evidence defaults on for consolidate analysis strategies and execute-merge (lifecycle, risk, warnings).",
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: true,
-      idempotentHint: false,
-      openWorldHint: true,
-    },
-    inputSchema: z.object({
-      cwd: projectParam,
-      strategy: z
-        .enum([
-          "detect-duplicates",
-          "find-clusters",
-          "suggest-merges",
-          "execute-merge",
-          "prune-superseded",
-          "dry-run",
-        ])
-        .describe(
-          "What to do: 'dry-run' = full analysis without changes, 'detect-duplicates' = find similar pairs, " +
-          "'find-clusters' = group by theme and relationships, 'suggest-merges' = actionable merge recommendations, " +
-          "'execute-merge' = perform a merge (requires mergePlan), 'prune-superseded' = delete notes marked as superseded. " +
-          "Use `evidence: true` on analysis strategies for trust/risk signals."
-        ),
-      mode: z
-        .enum(CONSOLIDATION_MODES)
-        .optional()
-        .describe("Override the project's default: 'supersedes' preserves history, 'delete' removes sources immediately"),
-      threshold: z
-        .number()
-        .min(0)
-        .max(1)
-        .optional()
-        .default(0.85)
-        .describe("Cosine similarity threshold for duplicate detection (0.85 default)"),
-      evidence: z
-        .boolean()
-        .optional()
-        .describe("Confidence signals for analysis strategies and execute-merge (lifecycle, risk, warnings). Default true for safety."),
-      mergePlan: z
-        .object({
-          sourceIds: z.array(z.string()).min(2).describe("Ids of notes to merge into one consolidated note"),
-          targetTitle: z.string().describe("Title for the consolidated note"),
-          content: z.string().max(100000, "Content must be at most 100,000 characters").optional().describe("Custom body for the consolidated note — distill durable knowledge rather than dumping all source content verbatim"),
-          description: z.string().optional().describe("Context explaining the consolidation rationale (stored in the note)"),
-          summary: z.string().optional().describe("Git commit summary only. Imperative mood, concise, and focused on why the change matters."),
-          tags: z.array(z.string()).optional().describe("Optional tags for later filtering. Use a small number of stable, meaningful tags."),
-        })
-        .optional()
-        .describe("Required for 'execute-merge' strategy. Get sourceIds from 'suggest-merges' output."),
-      allowProtectedBranch: z
-        .boolean()
-        .optional()
-        .describe(
-          "One-time override for protected branch checks. " +
-          "When true, consolidate can commit on a protected branch without changing project policy."
-        ),
-    }),
-    outputSchema: ConsolidateResultSchema,
-  },
-  async ({ cwd, strategy, mode, threshold, evidence = true, mergePlan, allowProtectedBranch = false }) => {
-    await ensureBranchSynced(cwd);
-
-    const project = await resolveProject(cwd);
-    if (!project && cwd) {
-      return projectNotFoundResponse(cwd);
-    }
-
-    // Gather notes from all vaults (project + main) for this project
-    const { entries } = await collectVisibleNotes(cwd, "all", undefined, "any");
-    const projectNotes = project
-      ? entries.filter((e) => e.note.project === project.id)
-      : entries.filter((e) => !e.note.project);
-
-    if (projectNotes.length === 0) {
-      return { content: [{ type: "text", text: "No memories found to consolidate." }], isError: true };
-    }
-
-    // Resolve project/default consolidation mode. Temporary-only merges may still
-    // resolve to delete later when a specific source set is known.
-    const policy = project ? await configStore.getProjectPolicy(project.id) : undefined;
-    const defaultConsolidationMode = resolveConsolidationMode(policy);
-
-      switch (strategy) {
-      case "detect-duplicates":
-        return detectDuplicates(projectNotes, threshold, project ?? undefined, evidence);
-
-      case "find-clusters":
-        return findClusters(projectNotes, project ?? undefined);
-
-      case "suggest-merges":
-        return suggestMerges(projectNotes, threshold, defaultConsolidationMode, project ?? undefined, mode, evidence);
-
-      case "execute-merge": {
-        if (!mergePlan) {
-          return { content: [{ type: "text", text: "execute-merge strategy requires a mergePlan with sourceIds and targetTitle." }], isError: true };
-        }
-        const mergeResult = await executeMerge(ctx, entries, mergePlan, defaultConsolidationMode, project ?? undefined, cwd, mode, policy, allowProtectedBranch, evidence);
-        invalidateActiveProjectCache();
-        return mergeResult;
-      }
-
-      case "prune-superseded": {
-        const pruneResult = await pruneSuperseded(ctx, projectNotes, mode ?? defaultConsolidationMode, project ?? undefined, cwd, policy, allowProtectedBranch);
-        invalidateActiveProjectCache();
-        return pruneResult;
-      }
-
-      case "dry-run":
-        return dryRunAll(projectNotes, threshold, defaultConsolidationMode, project ?? undefined, mode, evidence);
-
-      default: {
-        const _exhaustive: never = strategy;
-        return { content: [{ type: "text", text: `Unknown strategy: ${_exhaustive}` }], isError: true };
-      }
-    }
-  }
-);
-
-async function warnAboutPendingMigrationsOnStartup(): Promise<void> {
-  let totalPending = 0;
-  const details: string[] = [];
-
-  for (const vault of vaultManager.allKnownVaults()) {
-    const version = await readVaultSchemaVersion(vault.storage.vaultPath);
-    const pending = await migrator.getPendingMigrations(version);
-    if (pending.length === 0) {
-      continue;
-    }
-
-    totalPending += pending.length;
-    const label = vault.isProject ? "project" : "main";
-    details.push(
-      `${label} (${vault.storage.vaultPath}): ${pending.length} pending from schema ${version}`,
-    );
-  }
-
-  if (totalPending === 0) {
-    return;
-  }
-
-  console.error(
-    `[mnemonic] ${totalPending} pending migration(s) detected - run "mnemonic migrate --dry-run" to preview`,
-  );
-  for (const detail of details) {
-    console.error(`[mnemonic]   ${detail}`);
-  }
-}
+registerConsolidateTool(server, ctx);
 
 // ── mnemonic-workflow-hint prompt ─────────────────────────────────────────────
 server.registerPrompt(
@@ -2314,7 +805,9 @@ server.registerPrompt(
 );
 
 // ── start ─────────────────────────────────────────────────────────────────────
-await warnAboutPendingMigrationsOnStartup();
+import { warnAboutPendingMigrationsOnStartup } from "./startup.js";
+
+await warnAboutPendingMigrationsOnStartup(ctx);
 const transport = new StdioServerTransport();
 
 async function shutdown() {
